fix: data races

+++ usleep.c
@@ -6,7 +6,7 @@
 /*   By: macbook <macbook@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/22 16:13:43 by macbook           #+#    #+#             */
-/*   Updated: 2024/12/31 06:08:07 by macbook          ###   ########.fr       */
+/*   Updated: 2025/01/02 18:34:44 by macbook          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_usleep(useconds_t usec)
 {
-	long	start;
+	long long	start;
 
 	start = ft_get_time();
 	while ((ft_get_time() - start) < usec)
@@ -23,25 +23,19 @@ void	ft_usleep(useconds_t usec)
 	}
 }
 
+
 // void	ft_usleep(useconds_t usec)
 // {
 // 	long	start;
-// 	long	elapsed;
-// 	long	rem;
 
 // 	start = ft_get_time();
-// 	while (ft_get_time() - start < usec)
-// 	{
-// 		// if (simulation_finished(table))
-// 		// 	break ;
-// 		elapsed = ft_get_time() - start;
-// 		rem = usec - elapsed;
-// 		if (rem > 1e3)
-// 			usleep(rem / 2);
-// 		else
-// 		{
-// 			while (ft_get_time() - start < usec)
-// 				;
-// 		}
-// 	}
+// 	while ((ft_get_time() - start) < usec)
+// 		usleep(usec / 2);
+// }
+
+// void ft_usleep(useconds_t usec) {
+//     long start = ft_get_time(); // Get the starting time in microseconds
+//     while ((ft_get_time() - start) < usec) {
+//         usleep(10); // Small sleep interval to prevent busy-waiting
+//     }
 // }
